Update on "[JIT] during freezing, cast optional bias to half if weight is half"
On GPU, conv errors if not all its inputs have the same dtype.

In the case of autocasting during freezing, what we see is:
1) inputs to conv are casted to half
2) inputs to batchnorm are not casted, so many are still floats
3) we try to fold conv + batchnorm, by finding different weight and bias such that conv(input, new_weight, new_bias) is equivalent to the original conv -> batchnorm.

If conv previously had an optional bias, then during freezing we will temporarily create a zero-valued bias as a placeholder for conv_bias. We want to construct it to have the same dtype as the weight input to conv, to avoid errors on GPU.

Fixes #76578

[ghstack-poisoned]

+++ test/jit/test_freezing.py
@@ -1687,7 +1687,7 @@ class TestFrozenOptimizations(JitTestCase):
             def __init__(self, in_channels, out_channels, **kwargs):
                 super(ConvBN, self).__init__()
                 self.conv = torch.nn.Conv2d(in_channels, out_channels, bias=False, dtype=torch.half, **kwargs)
-                self.bn = torch.nn.BatchNorm2d(out_channels, eps=0.001, dtype=torch.half)
+                self.bn = torch.nn.BatchNorm2d(out_channels, eps=0.001, dtype=torch.float)
 
             def forward(self, x):
                 return self.bn(self.conv(x))
